moved extendDecorate tests to object tests
